+++ src/server/mod.rs
@@ -1,10 +1,17 @@
 use actix_files::Files;
-use actix_web::{dev::Server, middleware, web::JsonConfig, App, HttpServer, Result};
+use actix_web::{
+    dev::Server,
+    middleware,
+    web::{Data, JsonConfig},
+    App, HttpServer, Result,
+};
+use bigdecimal::FromPrimitive;
 use diesel::pg::PgConnection;
 use diesel::r2d2::{ConnectionManager, Pool};
 use std::io;
 use thiserror::Error;
 
+use crate::database::update_event_ids;
 use crate::kafka_consumer;
 
 mod route;
@@ -29,6 +36,7 @@ pub(crate) struct EtcdServer {
     pub(crate) docker_host_addr: String,
 }
 
+#[allow(clippy::mutex_atomic)]
 pub fn run(
     database_url: &str,
     reviewd_addr: &std::net::SocketAddr,
@@ -66,11 +74,24 @@ pub fn run(
             .expect("Unable to create Tokio runtime")
             .block_on(kafka_consumer::KafkaConfig::periodically_fetch_kafka_message(config));
     });
+    let max_event_id_num = std::env::var("MAX_EVENT_ID_NUM")
+        .ok()
+        .and_then(|v| v.parse::<usize>().ok())
+        .unwrap_or_else(|| 25);
+    log::info!(
+        "The maximum number of event_ids per cluster/outlier is {}",
+        max_event_id_num
+    );
+    if let (Ok(conn), Some(num)) = (pool.get(), FromPrimitive::from_usize(max_event_id_num)) {
+        let _ = update_event_ids(&conn, num);
+    }
+    let max_event_id_num = Data::new(std::sync::Mutex::new(max_event_id_num));
     let server = HttpServer::new(move || {
         App::new()
             .data(JsonConfig::default().limit(1_048_576))
             .data(pool.clone())
             .data(kafka_url.clone())
+            .app_data(max_event_id_num.clone())
             .data(etcd_server.clone())
             .configure(route::init_app)
             .service(Files::new("/", frontend_path.as_str()).index_file("index.html"))